requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
filelock==3.9.0
fonttools==4.38.0
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pipenv==2023.2.4
platformdirs==3.0.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
